Avances Lupita
-Se cambio un poco el diseño para corte y listado de caja, ademas se agrego la parte del diseño de facturacion.
-Se pueden ver los tickets sin facturacion
-se acoplaron las tablas al diseño web

+++ app/src/main/java/com/Danthop/bionet/model/FormaspagoModel.java
@@ -3,11 +3,11 @@ package com.Danthop.bionet.model;
 public class FormaspagoModel {
     private String id_pago;
     private String formapago;
-    private int importe;
+    private Double importe;
 
 
 
-    public FormaspagoModel(String id_pago, String formapago, int importe){
+    public FormaspagoModel(String id_pago, String formapago, Double importe){
         this.id_pago = id_pago;
         this.formapago = formapago;
         this.importe = importe;
@@ -16,13 +16,13 @@ public class FormaspagoModel {
 
     public String getid_pago() { return id_pago; }
     public String getformapago() { return  formapago; }
-    public int getImporte() { return importe; }
+    public Double getImporte() { return importe; }
 
 
 
     public void setid_pago(String id_pago) { this.id_pago = id_pago; }
     public void setformapago(String formapago) { this.formapago = formapago; }
-    public void setImporte(int importe) { this.importe = importe; }
+    public void setImporte(Double importe) { this.importe = importe; }
